requirements.txt
Django==1.11.10
djangorestframework==3.7.7
pytz==2018.3
